requirements.txt
aiofiles==0.7.0
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.1.0
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
cryptography==3.4.7
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.3
fastapi==0.65.2
fastapi-versioning==0.9.1
flit==3.4.0
flit-core==3.4.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httptools==0.2.0
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kaitaistruct==0.9
kombu==5.1.0
MarkupSafe==2.0.1
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dotenv==0.17.1
python-http-client==3.3.2
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
selenium-wire==4.3.0
sendgrid==6.7.1
six==1.16.0
starkbank-ecdsa==1.1.1
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
tomli-w==0.3.0
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.14.0
uvloop==0.15.2
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
wsproto==1.0.0
